refactor arena functions
the major change is adding the `cmt_pos` rather than having
a count of how many commits they were made.
this change has some regression which it doesn't support the
linked list arena, but the old implementation was already flawed
to begin with, for example it didn't actually free the node arena
when the arena was empty.